+++ src/lib/components/TreeView/TreeElement.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { onDestroy, onMount } from "svelte";
-    import { link, shared } from "../../shared.svelte";
-    import CheckBox from "../CheckBox.svelte";
+    import { link, shared } from "$lib/shared.svelte";
+    import CheckBox from "$lib/components/CheckBox.svelte";
     import TreeElement from "./TreeElement.svelte"
     let { slotId, unfolded = false, persistent = true, active = true }: { slotId: string, unfolded?: boolean, persistent?: boolean, active?: boolean } = $props()
 
@@ -44,7 +44,7 @@
     }
 
     // svelte-ignore state_referenced_locally
-    let slot = $state((await link.getSlot(slotId)).data)
+    let slot = $state.raw((await link.getSlot(slotId)).data)
 
     onDestroy(() => {
         shared.slotUpdate.delete(slotId)